{"types":"C","runs":12}

+++ Problem16.java
@@ -12,7 +12,7 @@ public class Problem16 extends ConsoleProgram {
 		int a = readInt ("Enter first number: ");
 		int b = readInt ("Enter second number: ");
 		int c = readInt ("Enter third number: ");
-		double harm =1 / (double)(1 / a + 1 / b + 1 / c);
+		double harm = 1 / (1 / a + 1 / b + 1 / c);
 		
 		println(harm);
 	}
